refactor api service and add create flight api handler
